Rename round numbers to more generic maths utility
