+++ api/databases/crud/symbols_crud.py
@@ -14,7 +14,8 @@ from sqlalchemy.orm import selectinload, QueryableAttribute
 from sqlalchemy.sql import delete
 from databases.utils import engine
 from tools.enum_definitions import QuoteAssets
-from tools.round_numbers import round_numbers
+from tools.maths import round_numbers
+
 
 class SymbolsCrud:
     """
@@ -338,26 +339,30 @@ class SymbolsCrud:
             if not existing_symbol:
                 price_filter = next(
                     (m for m in item["filters"] if m["filterType"] == "PRICE_FILTER"),
-                    None
+                    None,
                 )
                 quantity_filter = next(
-                    (m for m in item["filters"] if m["filterType"] == "LOT_SIZE"),
-                    None
+                    (m for m in item["filters"] if m["filterType"] == "LOT_SIZE"), None
                 )
                 min_notional_filter = next(
-                    (m for m in item["filters"] if m["filterType"] == "NOTIONAL"),
-                    None
+                    (m for m in item["filters"] if m["filterType"] == "NOTIONAL"), None
                 )
                 symbol = SymbolTable(
                     id=item["symbol"],
                     active=True,
-                    price_precision=round_numbers(price_filter["tickSize"]) if price_filter else 0,
-                    qty_precision=round_numbers(quantity_filter["stepSize"]) if quantity_filter else 0,
-                    min_notional=round_numbers(min_notional_filter["minNotional"]) if min_notional_filter else 0,
+                    price_precision=round_numbers(price_filter["tickSize"])
+                    if price_filter
+                    else 0,
+                    qty_precision=round_numbers(quantity_filter["stepSize"])
+                    if quantity_filter
+                    else 0,
+                    min_notional=round_numbers(min_notional_filter["minNotional"])
+                    if min_notional_filter
+                    else 0,
                     quote_asset=item["quoteAsset"],
                     base_asset=item["baseAsset"],
                     is_margin_trading_allowed=item["isMarginTradingAllowed"],
-                    asset_indices=[]
+                    asset_indices=[],
                 )
                 self.session.add(symbol)
         self.session.commit()
