+++ api/databases/utils.py
@@ -5,7 +5,7 @@ from sqlmodel import Session
 from time import time
 from typing import Annotated, Any
 from pydantic import BeforeValidator
-from tools.round_numbers import round_timestamp
+from tools.maths import round_timestamp
 
 # Load env vars for Alembic
 load_dotenv()
