+++ api/deals/abstractions/margin_deal_abstract.py
@@ -15,7 +15,7 @@ from tools.enum_definitions import (
     Strategy,
 )
 from tools.exceptions import BinanceErrors, MarginShortError
-from tools.round_numbers import (
+from tools.maths import (
     round_numbers,
     round_numbers_ceiling,
     round_timestamp,
