+++ api/deals/abstractions/spot_deal_abstract.py
@@ -14,7 +14,7 @@ from tools.enum_definitions import (
 from bots.models import BotModel, OrderModel, BotBase
 from deals.abstractions.factory import DealAbstract
 from deals.margin import MarginDeal
-from tools.round_numbers import round_numbers, round_timestamp
+from tools.maths import round_numbers, round_timestamp
 from urllib.error import HTTPError
