+++ api/deals/spot.py
@@ -2,7 +2,7 @@ from typing import Type, Union
 from databases.models.bot_table import BotTable, PaperTradingTable
 from tools.enum_definitions import QuoteAssets, DealType, Status, OrderSide, OrderStatus
 from bots.models import BotModel, OrderModel
-from tools.round_numbers import round_numbers, round_timestamp
+from tools.maths import round_numbers, round_timestamp
 from deals.abstractions.spot_deal_abstract import SpotDealAbstract
 from databases.crud.paper_trading_crud import PaperTradingTableCrud
