+++ api/tools/enum_definitions.py
@@ -160,4 +160,3 @@ class QuoteAssets(str, Enum):
     USDC = "USDC"
     BTC = "BTC"
     ETH = "ETH"
-
